requirements.txt
dvc==3.60.1
pandas==2.2.3
faiss-cpu==1.11.0.post1
fsspec==2025.3.2
huggingface-hub==0.34.3
joblib==1.5.1
langchain==0.3.27
langchain-classic==1.0.0
langchain-community==0.3.27
langchain-core==1.0.7
langchain-huggingface==0.3.1
langchain-openai==1.0.3
openai==2.8.1
pandas==2.2.3
pydantic==2.11.5
pymystem3==0.2.0
python-dotenv==1.1.0
rank-bm25==0.2.2
ruamel.yaml==0.18.14
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==5.0.0
# torch==2.3.0
tqdm==4.67.1